[Data Views for Rules] - adding unit tests (#24)

+++ x-pack/plugins/security_solution/server/lib/detection_engine/rules/bulk_action_edit.ts
@@ -24,9 +24,10 @@ export const deleteItemsFromArray = <T>(arr: T[], items: T[]): T[] => {
 
 export const applyBulkActionEditToRule = (
   existingRule: RuleAlertType,
-  action: BulkActionEditPayload
+  action: BulkActionEditPayload,
 ): RuleAlertType => {
   const rule = { ...existingRule, params: { ...existingRule.params } };
+
   switch (action.type) {
     // tags actions
     case BulkActionEditType.add_tags:
@@ -49,8 +50,11 @@ export const applyBulkActionEditToRule = (
         "Index patterns can't be added. Machine learning rule doesn't have index patterns property"
       );
 
-      rule.params.index = addItemsToArray(rule.params.index ?? [], action.value);
-      rule.params.dataViewId = undefined;
+      if (!(rule.params.dataViewId != null && rule.params.dataViewId.trim() !== '') || (rule.params.dataViewId != null && rule.params.dataViewId.trim() !== '' && action.overwriteDataViews)) {
+        rule.params.index = addItemsToArray(rule.params.index ?? [], action.value);
+        rule.params.dataViewId = undefined;
+      }
+
       break;
 
     case BulkActionEditType.delete_index_patterns:
@@ -59,13 +63,16 @@ export const applyBulkActionEditToRule = (
         "Index patterns can't be deleted. Machine learning rule doesn't have index patterns property"
       );
 
-      rule.params.index = deleteItemsFromArray(rule.params.index ?? [], action.value);
-      rule.params.dataViewId = undefined;
+      if (!(rule.params.dataViewId != null && rule.params.dataViewId.trim() !== '') || ((rule.params.dataViewId != null && rule.params.dataViewId.trim() !== '') && action.overwriteDataViews)) {
+        rule.params.index = deleteItemsFromArray(rule.params.index ?? [], action.value);
+        rule.params.dataViewId = undefined;
 
-      invariant(
-        rule.params.index.length !== 0,
-        "Can't delete all index patterns. At least one index pattern must be left"
-      );
+        invariant(
+          rule.params.index.length !== 0,
+          "Can't delete all index patterns. At least one index pattern must be left"
+        );
+      }
+
       break;
 
     case BulkActionEditType.set_index_patterns:
@@ -75,8 +82,11 @@ export const applyBulkActionEditToRule = (
       );
       invariant(action.value.length !== 0, "Index patterns can't be overwritten with empty list");
 
-      rule.params.index = action.value;
-      rule.params.dataViewId = undefined;
+      if (!(rule.params.dataViewId != null && rule.params.dataViewId.trim() !== '') || ((rule.params.dataViewId != null && rule.params.dataViewId.trim() !== '') && action.overwriteDataViews)) {
+        rule.params.index = action.value;
+        rule.params.dataViewId = undefined;
+      }
+
       break;
 
     // timeline actions
